retry: implement generic retrying function
This breaks out the retry logic into it's own separate function. This is
a cleaner and simpler organization strategy, and allows us to use the
retry logic on anything that returns a Result, instead of just the
client logic.
